Uploading the initial app
This is the initial application.  It is fully functional, other than a
concrete database for now.  I will be going through it and refactoring a
few things.  One major refactor will be adding Grails Services to better
align with best practices.  Another area will be adding a few more
tables.